feat: builders with validation for each object

+++ src/main/java/com/vorono4ka/swf/movieclips/MovieClipChild.java
@@ -1,4 +1,4 @@
 package com.vorono4ka.swf.movieclips;
 
-public record MovieClipChild(int id, byte blend, String name) {
+public record MovieClipChild(int id, int blend, String name) {
 }
